requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
argcomplete==2.0.0
astor==0.8.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
blessed==1.19.1
boto==2.49.0
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854164153/work
CacheControl==0.12.11
cachetools==5.2.0
cachy==0.3.0
causal-learn==0.1.3.1
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1625835301720/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1655906222726/work
cleo==1.0.0a5
click==8.1.3
cloudpickle==2.1.0
crashtest==0.3.1
crcmod==1.7
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1652967085355/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
dataclasses==0.6
debugpy @ file:///tmp/build/80754af9/debugpy_1637091426235/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dgl-cu111==0.6.1
docker-pycreds==0.4.0
dulwich==0.20.46
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908585034/work
et-xmlfile==1.1.0
fasteners==0.17.3
filelock==3.8.0
fonttools==4.34.4
frozendict==2.3.8
frozenlist==1.3.0
future==0.18.3
gcs-oauth2-boto-plugin==3.0
gitdb==4.0.9
GitPython==3.1.27
google-apitools==0.5.32
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-reauth==0.1.1
gpustat==1.0.0
GPUtil==1.4.0
graphviz==0.20.1
grpcio==1.47.0
gsutil==5.11
gym==0.25.1
gym-notices==0.0.7
html5lib==1.1
httplib2==0.20.4
huggingface-hub==0.8.1
hydra-core==1.1.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
idx2numpy==1.2.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel @ file:///tmp/build/80754af9/ipykernel_1646983213839/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1643818144432/work
jaraco.classes==3.2.2
jedi @ file:///tmp/build/80754af9/jedi_1644299024593/work
jeepney==0.8.0
Jinja2==3.1.2
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1650622202839/work
jupyter-core @ file:///opt/conda/conda-bld/jupyter_core_1651671229925/work
keyring==23.9.1
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
lockfile==0.12.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
monotonic==1.6
more-itertools==8.14.0
msgpack==1.0.4
multidict==6.0.2
munkres==1.1.4
natsort==8.3.1
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1649847911453/work
networkx==2.6.3
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
nvidia-ml-py==11.495.46
oauth2client==4.1.3
oauthlib==3.2.0
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
omegaconf==2.1.1
openpyxl==3.0.10
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
paddle-bfloat==0.1.7
paddlepaddle-gpu==2.3.1.post116
pandas==1.3.5
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathtools==0.1.2
patsy==0.5.3
pexpect==4.8.0
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1621287993571/work
pkginfo==1.8.3
pkgutil_resolve_name==1.3.10
poetry==1.2.0
poetry-core==1.1.0
poetry-plugin-export==1.0.7
porty==0.0.1
promise==2.3
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydot==1.4.2
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pylev==1.4.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
pyparsing==3.0.9
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1648857264451/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.2
pyu2f==0.1.5
PyYAML==6.0
pyzmq @ file:///opt/conda/conda-bld/pyzmq_1657724186960/work
qpth==0.0.15
regex==2022.8.17
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1656534056640/work
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
retry-decorator==1.1.1
rsa==4.7.2
sacremoses==0.0.53
SciencePlots==1.0.9
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1642601761909/work
scipy @ file:///tmp/build/80754af9/scipy_1641536880743/work
seaborn==0.11.2
SecretStorage==3.3.3
semver==3.0.0
sentry-sdk==1.9.5
shellingham==1.5.0
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
statsmodels==0.13.5
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor-whl==1.1.2
TFSnippet @ git+https://github.com/haowen-xu/tfsnippet.git@63adaf04d2ffff8dec299623627d55d4bacac598
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tokenizers==0.10.3
tomlkit==0.11.4
torch==1.7.0
torch-cluster==1.6.0
torch-geometric==2.1.0.post1
torch-scatter==2.0.9
torch-sparse==0.6.15
torch-spline-conv==1.2.1
torchaudio==0.12.0
torchvision==0.13.0
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm==4.64.0
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
transformers==4.11.3
typer==0.6.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1656706066251/work
umap==0.1.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1658789158161/work
wandb==0.12.10
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
Werkzeug==2.2.2
xlrd==1.2.0
yarl==1.7.2
yaspin==2.2.0
zipp==3.8.1
